[ADD] Get: Todos las salas de una sucursal.
Se arregla que las sucursales no se pueden eliminar si existen empleados en ella.
Primero hay que eliminar a los empleados relacionados a la sucursal.

+++ CineTec/CineTec/Context/CRUDContext.cs
@@ -70,6 +70,23 @@ namespace CineTec.Context
 
         }
 
+        public IList<Room> Get_all_rooms_of_a_branch(string cinema_name)
+        {
+            // Obtener todas las salas de la sucursal que coincide con el cinema_name ingresado.
+            var query = from b in Branches.Where(b => b.cinema_name == cinema_name)
+                        join room in Rooms
+                            on b.cinema_name equals room.branch_name
+                        select new { room };
+            var queryRoom = from t in query
+                            select t.room;
+            if (!queryRoom.Any())
+            {
+                return null;
+            }
+            IList<Room> myList = queryRoom.Cast<Room>().ToList();
+            return myList;
+        }
+
 
         public IList<Seat> Get_all_seats_of_a_room(string cinema_name, int id)
         {
@@ -91,7 +108,6 @@ namespace CineTec.Context
             IList<Seat> myList = querySeat.Cast<Seat>().ToList();
             return myList;
         }
-
 
 
         // Elimina las salas de una sucursal y luego elimina la sucursal misma.
@@ -125,21 +141,29 @@ namespace CineTec.Context
                 var queryEmp = from t in queryEmployees
                                select t.emp;
 
-                Seat[] s = querySeat.ToArray();
-                Room[] r = queryRoom.ToArray();
+                //Casteos
                 Employee[] e = queryEmp.ToArray();
+                if (e.Length == 0)
+                {
 
-                Seats.RemoveRange(s);
-                SaveChanges();
 
-                Rooms.RemoveRange(r);
-                SaveChanges();
+                    Seat[] s = querySeat.ToArray();
+                    Room[] r = queryRoom.ToArray();
 
-                Employees.RemoveRange(e);
-                SaveChanges();
+                    // Procede a borrar.
+                    Seats.RemoveRange(s);
+                    SaveChanges();
 
-                Branches.Remove(branch);
-                SaveChanges();
+                    Rooms.RemoveRange(r);
+                    SaveChanges();
+
+                    Branches.Remove(branch);
+                    SaveChanges();
+                }
+
+                ///////////// PONER UN AVISO AQUI DE QUE NO SE PUEDE BORRAR UNA SUCURSAL QUE TIENE EMPLEADOS
+                /// PRIMERO ES NECESARIO BORRAR LOS EMPLEADOS POR SI SOLOS.
+
             }
         }
